pyproject.toml
[project]
name = "FurManager"
version = "0.1.0"
description = "FurManager is a Telegram bot for channels autoposting and management."
authors = [
    { name = "Andrii Borychevskyi", email = "a.borychevskyi@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "dishka>=1.7.2",
    "greenlet>=3.3.0",
    "loguru>=0.7.3",
    "orjson>=3.11.5",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
]


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
mypy = "^1.16.0"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["alembic", "venv", ".venv", "tests", "modules"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
check_untyped_defs = true
explicit_package_bases = true
mypy_path = "."
exclude = "(^alembic/|^venv/|^\\.venv/|^tests/|^modules/)"
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["untyped_package.*"]
follow_untyped_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
